Resync with brotli-1.0.9 dist

+++ c/enc/bit_cost.h
@@ -45,7 +45,7 @@ static BROTLI_INLINE double BitsEntropy(
     const uint32_t* population, size_t size) {
   size_t sum;
   double retval = ShannonEntropy(population, size, &sum);
-  if (retval < (double)sum) {
+  if (retval < sum) {
     /* At least one bit per literal is needed. */
     retval = (double)sum;
   }
